requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.2.3
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
aiogram==3.4.1
aiohttp==3.9.1
playwright==1.49.1